Fix argument expansion when invoking composer
`$*` is problematic because it will blow up if any of the arguments passed with quotes contains spaces. `"$*"` is better, but `"$@"` won't explode if `IFS` is set to something strange, so that's the best solution. See e.g. http://stackoverflow.com/questions/448407/bash-script-to-receive-and-repass-quoted-parameters.

To reproduce, try running `composer config vendor-dir "some path"`, it will fail, but with this change, it works.